[chore]Print allocations, print unallocations, print rooms adjusted to work

+++ amityapp/app/amity.py
@@ -1,4 +1,6 @@
 import random
+import os
+import os.path
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
@@ -15,8 +17,12 @@ class Amity(object):
         self.people = []
         self.offices = []
         self.offices_with_members = []
+        self.office_members = {}
+        self.unallocated_offices = []
         self.living_spaces = []
         self.living_spaces_with_mem = []
+        self.living_mem = {}
+        self.unallocated_living = []
         self.rooms = []
         self.office_allocations = {}
         self.living_space_allocations = {}
@@ -24,7 +30,8 @@ class Amity(object):
     def create_room(self, room_type, room_name):
         """function to create a room."""
         try:
-            if not room_type.title() == "OFFICE" and not room_type.title() == "LIVINGSPACE":
+            if not room_type.title() == "Office" and not room_type.title() ==\
+                    "Livingspace":
                 raise NameError
             if room_name in self.rooms:
                 print("{} already exists". format(room_name))
@@ -79,13 +86,19 @@ class Amity(object):
                              len(office.office_members) < office.max_capacity]
         if available_offices:
             available_office = random.choice(available_offices)
-            available_office.office_members.append(name)
-            self.offices_with_members.append(available_office)
-            print("{} added to {} office".format(name.upper(), available_office.room_name.upper()))
-            print("Successful!")
+            for office in self.offices:
+                if name in office.office_members:
+                    break
+                else:
+                    available_office.office_members.append(name)
+                    self.offices_with_members.append(available_office)
+                    print("{} added to {} office".format(
+                        name.upper(), available_office.room_name.upper()))
+                    print("Successful!")
 
         else:
-            print("There are no offices available")
+            print("There are no offices available. {} added to unallocated list".format(name))
+            self.unallocated_offices.append(name)
 
     def allocate_living_space(self, name, wants_accommodation="N"):
         available_living_spaces = [living_space for living_space in
@@ -95,13 +108,17 @@ class Amity(object):
         if available_living_spaces:
             available_living_space = random.choice(available_living_spaces)
             if wants_accommodation == "Y":
-                available_living_space.living_space_members.append(name)
-                self.living_spaces_with_mem.append(available_living_space)
-                print("{} added to {} living space".format(name,
-                                                           available_living_space.room_name))
-                print("Successful!")
+                for living_space in self.living_spaces:
+                    if name not in living_space.living_space_members:
+                        available_living_space.living_space_members.append(name)
+                        self.living_spaces_with_mem.append(available_living_space)
+                        self.living_mem[available_living_space] = available_living_space.living_space_members
+                        print("{} added to {} living space".format(name,
+                                                                   available_living_space.room_name))
+                        print("Successful!")
         else:
-            print("There are no living spaces available")
+            print("There are no living spaces available. {} added to unallocated list")
+            self.unallocated_living.append(name)
 
     def reallocate_person(self, first_name, last_name, new_room_name):
         """function to re-allocate person to a new room"""
@@ -112,54 +129,70 @@ class Amity(object):
                                    self.living_spaces if
                                    len(living_space.living_space_members)
                                    < living_space.max_capacity]
+
         if full_name not in self.people:
             print("Person does not exist")
+        elif new_room_name not in self.rooms:
+            print("Room does not exist")
         else:
-            if new_room_name not in self.rooms:
-                print("Room does not exist")
-            else:
-                for office_room in self.offices_with_members:
-                    for name in office_room.office_members:
-                        for available_office in available_offices:
-                            if full_name == name:
-                                if new_room_name == available_office.room_name:
-                                    available_office.office_members.append(full_name)
-                                    office_room.office_members.remove(name)
-
-                    print("{} successfully reallocated to {}".format(full_name, new_room_name))
-
-                for living_space_room in self.living_spaces_with_mem:
-                    for name in living_space_room.living_space_members:
-                        for available_living_space in available_living_spaces:
-                            if full_name == name:
-                                if new_room_name == available_living_space.room_name:
-                                    available_living_space.living_space_members.append(full_name)
-                                    living_space_room.living_space_members.remove(name)
-                    print("{} successfully reallocated to {}".format(full_name,
-                                                                     new_room_name))
-# test living space
-# add exceptions
+            for available_living_space in available_living_spaces:
+                if new_room_name == available_living_space.room_name:
+                    print("Cannot reallocate from office to living space")
+                else:
+                    for keys, values in self.office_members.items():
+                        if full_name in values:
+                            if new_room_name == keys.room_name:
+                                print("Cannot reallocate to the same room")
+                            else:
+                                keys.office_members.remove(full_name)
+                                for available_office in available_offices:
+                                    if new_room_name == available_office.room_name:
+                                        available_office.office_members.append(full_name)
+                                        print("{} allocated to {} office".format(full_name,
+                                                                                 new_room_name))
+            for available_office in available_offices:
+                if new_room_name == available_office.room_name:
+                    pass
+                else:
+                    for keys, values in self.living_mem.items():
+                        if full_name in values:
+                            if new_room_name == keys.room_name:
+                                print("Cannot reallocate to the same room")
+                            else:
+                                keys.living_space_members.remove(full_name)
+                                for available_living_space in available_living_spaces:
+                                    if new_room_name == available_living_space.room_name:
+                                        available_living_space.living_space_members.append(
+                                            full_name)
+                                        print("{} allocated to {} living space".format(full_name,
+                                                                                       new_room_name))
 
     def load_people(self, text_file):
         """adds people from a text file"""
-        with open(text_file) as txt_file:
-            file_name = txt_file.readlines()
-            for sentence in file_name:
-                words = sentence.split()
-                if len(words) == 3:
-                    first_name = words[0].title()
-                    last_name = words[1].title()
-                    designation = words[2].title()
-                    wants_accommodation = "N"
-                    self.add_person(first_name, last_name, designation,
-                                    wants_accommodation)
-                elif len(words) == 4:
-                    first_name = words[0].title()
-                    last_name = words[1].title()
-                    designation = words[2].title()
-                    wants_accommodation = words[3].title()
-                    self.add_person(first_name, last_name, designation,
-                                    wants_accommodation)
+        if os.path.isfile(text_file) is False:
+            print("File does not exist")
+        else:
+            with open(text_file, 'r') as txt_file:
+                if (os.stat(text_file).st_size == 0):
+                    print('File is empty!')
+                else:
+                    file_name = txt_file.readlines()
+                    for sentence in file_name:
+                        words = sentence.split()
+                        if len(words) == 3:
+                            first_name = words[0].title()
+                            last_name = words[1].title()
+                            designation = words[2].title()
+                            wants_accommodation = "N"
+                            self.add_person(first_name, last_name, designation,
+                                            wants_accommodation)
+                        elif len(words) == 4:
+                            first_name = words[0].title()
+                            last_name = words[1].title()
+                            designation = words[2].title()
+                            wants_accommodation = words[3].title()
+                            self.add_person(first_name, last_name, designation,
+                                            wants_accommodation)
 
     def load_state(self, db_name='amity.db'):
         """loads data from database into application"""
@@ -209,39 +242,55 @@ class Amity(object):
 
     def print_allocations(self, file_name):
         """prints a list of allocations"""
-        file_open = open(file_name, 'w')
-        for office in self.offices:
-            if len(office.office_members) > 0:
-                print(office.room_name)
-                print(office.office_members)
-                file_open.write("{}: {}\n".format(office.room_name, office.office_members))
-        for living_space in self.living_spaces:
-            if len(living_space.living_space_members) > 0:
-                print(living_space.room_name)
-                print(living_space.living_space_members)
-                file_open.write("{}: {}\n".format(living_space.room_name,
-                                                  living_space.living_space_members))
+        for key_off, value_off in self.office_members.items():
+            print(key_off.room_name)
+            print("-------------------------")
+            print(", ".join(value_off))
+            if os.path.isfile(file_name) is False:
+                print("File does not exist")
+            else:
+                file_open = open(file_name, "w")
+                file_open.write("{}: {}\n".format(key_off.room_name,
+                                                  ", ".join(value_off)))
+                file_open.close()
+        for key_liv, value_liv in self.living_mem.items():
+            print(key_liv.room_name)
+            print("-------------------------")
+            print(", ".join(value_liv))
+            if os.path.isfile(file_name) is False:
+                print("File does not exist")
+            else:
+                file_open = open(file_name, "w")
+                file_open.write("{}: {}\n".format(key_liv.room_name,
+                                                  ", ".join(value_liv)))
+                file_open.close()
 
-    def print_unallocated(self):
+    def print_unallocated(self, file_name):
         """prints a list of unallocated people"""
-        for fellow in self.fellows:
-            fellow_name = fellow.first_name + fellow.last_name
-            for staff_p in self.staff:
-                staff_name = staff_p.first_name + staff_p.last_name
-                for office in self.offices:
-                    for living_space in self.living_spaces:
-                        if fellow.wants_accommodation == "Y" and fellow_name\
-                                not in living_space.living_space_members:
-                            print(fellow_name)
-                    if fellow_name and staff_name not in office.office_members:
-                        print(fellow_name)
-                        print(staff_name)
+        for office in self.unallocated_offices:
+            print("=====awaiting allocation to office ======")
+            print(office)
+            if os.path.isfile(file_name) is False:
+                print("File does not exist")
+            else:
+                file_open = open(file_name, "w")
+                file_open.write("{}: {}\n".format(office))
+                file_open.close()
+        for living_space in self.unallocated_living:
+            print("=====awaiting allocatopn to living space =====")
+            print(living_space)
+            if os.path.isfile(file_name) is False:
+                print("File does not exist")
+            else:
+                file_open = open(file_name, "w")
+                file_open.write("{}: {}\n".format(living_space))
+                file_open.close()
 
     def print_room(self, room_name):
         """prints names of people in room"""
         for office in self.offices:
-            print(office.room_name)
-            print(office.office_members)
+            if room_name == office.room_name:
+                print(", ".join(office.office_members))
         for living_space in self.living_spaces:
-            print(living_space.room_name)
-            print(living_space.living_space_members)
+            if room_name == living_space.room_name:
+                print(", ".join(living_space.living_space_members))
